+++ checkpoint/orbax/checkpoint/experimental/v1/_src/layout/numpy_layout.py
@@ -154,6 +154,9 @@ class NumpyLayout(CheckpointLayout):
           f'Failed to validate {path} as NumPy checkpoint: {e}'
       ) from e
 
+  async def get_checkpointable_names(self, path: Path) -> list[str]:
+    return [checkpoint_layout.PYTREE_CHECKPOINTABLE_KEY]
+
   async def validate_pytree(
       self, path: Path, checkpointable_name: str | None
   ) -> None: